Moves Footer to events folder and rename it
