+++ components/events/EventFooter.vue
